Rather than retrieving all values from the database after an object is flushed, retrieve only the primary key, if necessary.  All other values are set to AutoReload, and will only be retrieved once the one of them is touched.
This should improve the performance of Storm-based applications
significantly, as inserted values are only really retrieved if
needed.  In cases where the primary key is preset explicitly or
by the backend during the insertion, this will prevent the second
query entirely.  In other cases, it will mean a fast index lookup.