Remove PostgresConnection.raw_execute workaround for lack of Unicode queries.
This workaround was added in 2007, and presumably there was some
psycopg2 problem at the time that required it.  However, I've tested as
far back as psycopg2 2.0.10 (the oldest version still available on PyPI)
and there's no sign of it; I've also successfully run the full Launchpad
test suite without this workaround.  I think this must have been a
psycopg2 bug that was fixed a long time ago, and it now seems to be safe
to drop the workaround.