Add Python 3 support to tests.mocker.SpecChecker.
inspect.getargspec is deprecated on Python 3, and in any case causes us
problems because it includes the bound first argument (self or similar) for
bound methods.  The modern inspect.signature API is for the most part much
easier to use and doesn't have this problem.

There is one wrinkle which requires special handling: method descriptors
don't have the first argument already bound as far as their signature is
concerned, but for the purposes of a mock spec we want to skip it anyway.