Fix C version of Compile.__call__(join=) on Python 3.
On Python 3, an "S" argument only takes bytes, but this parameter should
only accept text.  (It still accepts either bytes or text on Python 2.)

The reason this wasn't caught in tests was a little subtle.  tox installs
the C extension in a virtualenv, as you might expect; but, when running the
test program, the storm package is imported from the same directory as the
test program, and so it never actually finds any C extensions.  The simplest
workaround is to move the test program to a subdirectory, where no storm
package exists and so it will be properly imported from the virtualenv.