Fix list() on security-proxied ResultSets on Python 3.
On Python 3, list() calls len() internally to preallocate the correct
amount of space if possible.  If __len__ is unimplemented, then it falls
back to guessing, but ForbiddenAttribute confuses it and results in
failing to create the list.  Allowing __len__ through the security
proxy, even though it's unimplemented, avoids this.

(We don't want to implement ResultSet.__len__ in terms of
ResultSet.count because it would make it too easy to write code that
implicitly issued extra SQL queries.  On the other hand,
EmptyResultSet.__len__ can trivially be implemented, so do so here.)