Avoid traceback reference cycles when wrapping exceptions.
`wrap_exceptions` stored the original exception's traceback in a local
variable, but the traceback in turn contains a reference to that frame,
creating a reference cycle.  The situation is worse on Python 3, because the
original exception also has the associated traceback as a `__traceback__`
attribute, producing more complicated reference cycles.  Delete these local
variables before returning to avoid leaking memory over time.