Clarify exception for a property with allow_none=False and default=None.
Creating a property with `allow_none=False, default=None` doesn't make
sense, but it's sometimes easy to set things up that way by accident
during refactoring.  Storm immediately raises `NoneError` when creating
the variable in that case.  However, it's not obvious that the exception
is due to the bad default, and I spent some time with a colleague today
debugging a situation that turned out to be due to this.

Change the exception message from "None isn't acceptable as a value" to
"None isn't acceptable as a default value" in this case, to clue future
travellers into the fact that the problem relates to the default.